requirements.txt
tabula-py>=2.7.0
pandas>=1.5.0